+++ custom_components/thz/climate.py
@@ -403,7 +403,7 @@ async def async_setup_entry(
         )
 
     if entities:
-        assign_subdevices(entities, config_entry.data)
+        assign_subdevices(entities, config_entry)
         async_add_entities(entities)
         _LOGGER.debug("Created %d climate entities", len(entities))
     _async_remove_dhw_climate(hass, config_entry)